backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.5.0
websockets==12.0
httpx==0.27.0
anthropic==0.39.0
openai==1.51.0
google-generativeai==0.8.0
pytest==8.3.0
pytest-asyncio==0.24.0
